added team remove validation

+++ src/controllers/teams.controller.ts
@@ -17,12 +17,14 @@ const RemoveMemberInOrganization = asyncHandler(
     if (!userId || !organization) {
       return new ApiResponse(400, null, "All Fields are required");
     }
+    if (req.user?.userId.toString() === userId?.toString()) throw new ApiError("Can't remove itself", 400);
     const checkUser = await Teams.findOne({
       userId: new mongoose.Types.ObjectId(req.user?.userId),
       Organization: new mongoose.Types.ObjectId(organization as string),
     });
     if (!checkUser) throw new ApiError("Unauthorised Member Found", 401);
     if (checkUser.role == "owner") {
+
       const removeMember = await Teams.deleteOne({
         userId: new mongoose.Types.ObjectId(userId!),
         Organization: checkUser?.Organization,
